requirements.txt
allure-pytest
appium-python-client
pytest
pytest-testconfig
requests
selenium
PyYaml
httprunner